requirement.txt
alembic==1.13.2
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.4
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.2
python-decouple==3.8
pytz==2024.1
referencing==0.35.1
rpds-py==0.20.0
SQLAlchemy==2.0.32
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.4
